requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
django==3.2.16
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-filter==22.1
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.25.1
drf-standardized-errors==0.12.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.2
jsonschema==4.17.3
markupsafe==2.1.1
mysqlclient==2.1.1
numpy==1.24.1
oauthlib==3.2.2
opencv-contrib-python==4.5.5.62
pillow==9.4.0
pycparser==2.21
pyjwt==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.7.1
pyyaml==6.0
requests==2.28.2
requests-oauthlib==1.3.1
setuptools==66.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
